refactor: async I/O as prerequisite of using S3 (#234)
